Adição DER, Script DB e Collections postman junto com novas classes controller e tentando o calculo de ticket

+++ src/main/java/br/com/estacionamento/estacionAqui/repository/EstacionamentoRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import br.com.estacionamento.estacionAqui.model.Estacionamento;
 
 @Repository
-public interface EstacionamentoRepository extends JpaRepository<Estacionamento, Integer>{
+public interface EstacionamentoRepository extends JpaRepository<Estacionamento, Long>{
 
 }
